+++ test/membrane/testing/dynamic_source_test.exs
@@ -39,7 +39,6 @@ defmodule Membrane.Testing.DynamicSourceTest do
             ]
       )
 
-    assert_pipeline_play(pipeline)
     assert_sink_buffer(pipeline, :sink_1, %Buffer{payload: 'a'})
     assert_sink_buffer(pipeline, :sink_1, %Buffer{payload: 'b'})
     assert_sink_buffer(pipeline, :sink_1, %Buffer{payload: 'c'})
@@ -63,7 +62,6 @@ defmodule Membrane.Testing.DynamicSourceTest do
             ]
       )
 
-    assert_pipeline_play(pipeline)
     assert_sink_buffer(pipeline, :sink_1, %Buffer{payload: <<0::16>>})
     assert_sink_buffer(pipeline, :sink_1, %Buffer{payload: <<1::16>>})
     assert_sink_buffer(pipeline, :sink_1, %Buffer{payload: <<2::16>>})